move hero to homepage and tests to homepage test file
